+++ docs/CHANGELOG.md
@@ -5,13 +5,15 @@
 Released: ???
 
 Features:
+- Added option to set a keybind in Minecraft's Controls menu to pause all active overlays on button pressed. Default button is PAUSE.
+  - The effect is the same as pressing [Click to pause] line on each active overlay.
 - Added option to auto-share found hotspots to the selected chat [disabled by default].
 - Announce RARE DROP! on Caster VI book dropped.
 - Do not show Abandoned Quarry tracker when in Trophy Hunter armor (as it's often used to just catch treasure).
 - Show Sea creature HP overlay and render boxes even if no fishing rod in hotbar.
 - Changed different overlays to hide them when player is not fishing.
   - Before this change, overlays were visible if player had a fishing rod in hotbar (which is a case for many other activities other than fishing).
-  - Now the overlays appear after player starts fishing in water/lava, and are hidden after 10 minutes of inactivity or after swapping server.
+  - Now the overlays appear after player starts fishing in water/lava, and hide after 10 minutes of inactivity / after swapping server.
   - Also, overlays with timers are not paused now when taking off a rod from hotbar.
   - The following overlays affected:
     - Fishing Profit tracker
@@ -27,6 +29,7 @@ Bugfixes:
 
 Other:
 - Refactored different functionality to not execute any checks when setting is disabled / when being in a wrong world:
+  - Fishing profit tracker
   - Sea creatures count overlay & alert
   - Worm membrane profit tracker